+++ src/semantic/default.ts
@@ -21,20 +21,14 @@ export function getDefaultSemantic(palette: Palette) {
     "property.defaultLibrary:javascriptreact": `${palette.purple}`,
     // }}}
     // TypeScript {{{
-    "class:typescript": `${palette.aqua}`,
-    "interface:typescript": `${palette.aqua}`,
     "enum:typescript": `${palette.purple}`,
     "enumMember:typescript": `${palette.blue}`,
-    "namespace:typescript": `${palette.purple}`,
     "variable.defaultLibrary:typescript": `${palette.purple}`,
     "property.defaultLibrary:typescript": `${palette.purple}`,
     // }}}
     // TypeScript React {{{
-    "class:typescriptreact": `${palette.aqua}`,
-    "interface:typescriptreact": `${palette.aqua}`,
     "enum:typescriptreact": `${palette.purple}`,
     "enumMember:typescriptreact": `${palette.blue}`,
-    "namespace:typescriptreact": `${palette.purple}`,
     "variable.defaultLibrary:typescriptreact": `${palette.purple}`,
     "property.defaultLibrary:typescriptreact": `${palette.purple}`,
     // }}}